server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-multipart
openai>=1.40.0
afinn
